+++ src/types/Invariant.ts
@@ -1,0 +1,1 @@
+export type InvariantFnType<T = unknown> = (value: T) => null | string | undefined;